requirements.txt
alembic==1.7.5
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
distlib==0.3.4
docopt==0.6.2
filelock==3.4.2
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
pipreqs==0.4.11
platformdirs==2.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.8
virtualenv==20.13.0
Werkzeug==2.0.2
yarg==0.1.9
